lambda/src/compiler/requirements.txt
humanfriendly
more_itertools
pyyaml
voluptuous
